% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaatech_xrf_corescanner_baxil.R
\name{batch_process_avaatech_baxil_csv}
\alias{batch_process_avaatech_baxil_csv}
\title{Batch process Avaatech XRF-CS bAXIL csv files}
\usage{
batch_process_avaatech_baxil_csv(
  avaa_baxil_csv_path,
  cut_off_set = TRUE,
  delim = ",",
  show_col_types = FALSE,
  xrf_data_start = 8,
  ...
)
}
\arguments{
\item{avaa_baxil_csv_path}{Absolute or relative path to the csv file}

\item{cut_off_set}{Should the offset at the top of the measurement be substracted (default is TRUE)? The offset caused by material that was used to close the sediment core will be substracted.}

\item{delim}{What csv file delimiter should be used (default is ",")?}

\item{show_col_types}{Show column types that were initially guessed by readr::read_delim() when reading the csv file. May be useful for debugging (default is FALSE)}

\item{xrf_data_start}{Number of the column in which the XRF data starts (e.g. `Sn-Ka Area`).}

\item{...}{Further optional arguments that are passed on to read_delim()}
}
\value{
A dataframe/tibble containing tidy Avaatech XRF-CS data. Multiple XRF-CS datasets are rowbound to one big table.
}
\description{
This function tries to validate and parse Avaatech XRF-Corescanner files created with bAXIL Batch (and saved as csv).
}
\details{
The function expects the following fields and conditions in the csv file:
* The first field in the file must be called `Spectrum`. The CoreID, Voltage and Current are extracted from that field.
* The sample depth must be in a field called `Sample_ID`.
* The XRF data must be in a wide form like `" Sn-Ka Chi2" ," Sn-Ka Area" ," Sn-Ka AreaStd" ," Sn-Ka cps" ," Sn-Ka cpsStd"`. Element and Parameter need to be separated with an underline. The function expects to split the names into 4 parts which are Element, Absorption Line, Scattering Type (Inc/Coh e.g. for Rh) and Parameter.

WARNING: This function does not filter out ambiguous data (e.g. same Element measured for different runs/voltage)!
}
