% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beckman_grainsizer.R
\name{batch_process_beckman_grainsizer}
\alias{batch_process_beckman_grainsizer}
\title{Batch process Beckman Coulter LS13 320 Grainsizer txt files}
\usage{
batch_process_beckman_grainsizer(beckman_gs_txt_files, calculate_depths = TRUE)
}
\arguments{
\item{beckman_gs_txt_files}{List or vector of txt files that should be parsed}

\item{calculate_depths}{Logical. Should Depths be calculated from the SampleID field? Defaults to `TRUE`}
}
\value{
A list of 4 dataframes for Parameters, Statistics, Interpolations and Channel data respectively, reflecting the structure of our txt export files that have 4 sections. The dataframes are rowbound for each sample/file.
}
\description{
Batch process Beckman Coulter LS13 320 Grainsizer txt files
}
