% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{carrrotter_example_files}
\alias{carrrotter_example_files}
\title{Get path to carrrotter example}
\usage{
carrrotter_example_files(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
carrrotter comes bundled with a few correct bAXIL csv files inside the `inst/extdata`
directory. This function takes the name of a file or multiple files and returns the path.
}
\examples{
carrrotter_example_files()
carrrotter_example_files(c("MI18-L8-A_10kV_raw.csv", "MI18-L8-B_10kV_raw.csv"))
}
