% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beckman_grainsizer.R
\name{make_grainsize_category_table}
\alias{make_grainsize_category_table}
\title{Create a human readable table of grainsize categories}
\usage{
make_grainsize_category_table(gsdata = NULL)
}
\arguments{
\item{gsdata}{Expects the output of `batch_process_beckman_grainsizer` as argument.}
}
\value{
Returns a human-readable dataframe/tibble that can be used to write to file
}
\description{
This function creates a human readable table of grainsize categories, differentiating between clay, silt sand.
}
