% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beckman_grainsizer.R
\name{parse_beckman_grainsizer_txt}
\alias{parse_beckman_grainsizer_txt}
\title{Parse a Beckman Coulter LS13 320 Grainsizer txt file}
\usage{
parse_beckman_grainsizer_txt(beckman_gs_txt_path, calculate_depths = TRUE)
}
\arguments{
\item{beckman_gs_txt_path}{Path to the export txt file}

\item{calculate_depths}{Logical. Should Depths be calculated from the SampleID field? Defaults to `TRUE`}
}
\value{
A list of 4 dataframes for Parameters, Statistics, Interpolations and Channel data respectively, reflecting the structure of our txt export files that have 4 sections.
}
\description{
This function parses the export files produced by the LS13 320 software in conjunction with the SOP used by the SURF Sedi Group at Eawag. It is most likely specific to the settings chosen for our group but can be adapted to other settings.
}
\details{
This function is not exported. Use `batch_process_beckman_grainsizer` instead.
}
