% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS13320.R
\name{read_LS13320}
\alias{read_LS13320}
\title{Parse export files from LS13 320}
\usage{
read_LS13320(path = ".", pattern = ".+\\\\.\\\\$ls\\\\.txt")
}
\arguments{
\item{pattern}{Regex file pattern to filter files in input directory. Default pattern is ".+\\.\\$ls\\.txt"}
}
\value{
A list of dataframes with one dataframe per paragraph in the well-structured export text files.
}
\description{
This function handles the export files generated by a Beckman-Coulter LS13 320 Particle Size Analyser and parses them into tidy dataframes.
}
