% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS13320.R
\name{read_LS13320_export}
\alias{read_LS13320_export}
\title{Parse export files from Beckman LS13 320}
\usage{
read_LS13320_export(path, pattern = ".+\\\\.\\\\$ls\\\\.txt")
}
\arguments{
\item{path}{The path to the export files.}

\item{pattern}{Regular expression pattern to filter files in input directory. Default pattern is '.+\\.\\$ls\\.txt' and will filter *.$ls.txt files.}
}
\value{
A list of dataframes with one dataframe per paragraph contained in the well-structured LS13 320 export files.
}
\description{
This function handles the export files generated by a Beckman-Coulter LS13 320 Particle Size Analyser and parses them into tidy dataframes. The function will return the parameter paragraph, the descriptive statistics paragraph, the interpolated particle size paragraph and the raw channel data paragraph or any subset of these if existing.
}
