% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaatechxrf.R
\name{read_xrfcs_avaatech_baxil}
\alias{read_xrfcs_avaatech_baxil}
\title{Parse Avaatech bAXIL Batch XRF-Corescanner files}
\usage{
read_xrfcs_avaatech_baxil(file)
}
\arguments{
\item{file}{Character vector with filenames (and full paths) of bAXIL Batch *.csv-files}
}
\value{
Tidied bAXIL XRF data, returned as a dataframe
}
\description{
This function takes a string with the path to your bAXIL Batch csv-files, reads
them and turns them into tidied dataframes.
}
\details{
Currently the function automatically removes elements with coherent (Coh) and
incoherent (Inc) traces and chooses the less noisy element (smaller relative
standard deviation), if present for different voltages. By default, the function removes the top offset of the data. Additionally, repeated measures are removed by default.
}
